Extract agent-task review command helpers
