+++ src/commands/mod.rs
@@ -167,6 +167,7 @@ pub fn finalize_set_spec(
 
 pub mod agent_task;
 pub(crate) mod agent_task_dispatch;
+pub(crate) mod agent_task_review;
 pub mod api;
 pub mod audit;
 pub mod audit_baseline;
